# docs/openapi_client.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Caser Client API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/cases/{case_id}/work-logs": {
      "post": {
        "tags": [
          "work-logs"
        ],
        "summary": "Create Work Log",
        "description": "Manual entry — the case must be one this lawyer is currently assigned\nto, and not closed (CLAUDE.md: a closed case blocks new WorkLogs the same\nway it blocks new Documents).",
        "operationId": "create_work_log_cases__case_id__work_logs_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkLogRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkLogResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "work-logs"
        ],
        "summary": "List Work Logs",
        "operationId": "list_work_logs_cases__case_id__work_logs_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_WorkLogResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}/work-logs/{work_log_id}": {
      "patch": {
        "tags": [
          "work-logs"
        ],
        "summary": "Update Work Log",
        "description": "Any lawyer assigned to the case can edit any entry on it, not just\ntheir own (CLAUDE.md: same broad, collaborative authority as Documents).\nLocked entirely once the case is closed — same cutoff as creating new\nentries, one consistent \"closed means frozen\" rule. Every edit is logged\nin AuditLogs (who changed which entry, when — not before/after values).",
        "operationId": "update_work_log_cases__case_id__work_logs__work_log_id__patch",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          },
          {
            "name": "work_log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Work Log Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkLogRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkLogResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "work-logs"
        ],
        "summary": "Delete Work Log",
        "operationId": "delete_work_log_cases__case_id__work_logs__work_log_id__delete",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          },
          {
            "name": "work_log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Work Log Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateWorkLogRequest": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "hours": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 24.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Hours"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "date",
          "hours"
        ],
        "title": "CreateWorkLogRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_WorkLogResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkLogResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[WorkLogResponse]"
      },
...
      "UpdateWorkLogRequest": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "hours": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 24.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Hours"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "date",
          "hours"
        ],
        "title": "UpdateWorkLogRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkLogResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "case_id": {
            "type": "integer",
            "title": "Case Id"
          },
          "lawyer_id": {
            "type": "integer",
            "title": "Lawyer Id"
          },
          "lawyer_name": {
            "type": "string",
            "title": "Lawyer Name"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "hours": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Hours"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "source": {
            "$ref": "#/components/schemas/WorkLogSource"
          }
        },
        "type": "object",
        "required": [
          "id",
          "case_id",
          "lawyer_id",
          "lawyer_name",
          "date",
          "hours",
          "source"
        ],
        "title": "WorkLogResponse"
      },
      "WorkLogSource": {
        "type": "string",
        "enum": [
          "manual",
          "excel_import"
        ],
        "title": "WorkLogSource"
      }
    }
  }
}